move the timeout test to a separate file.
