+++ loggers_test.go
@@ -7,7 +7,6 @@ import (
 	"strconv"
 	"strings"
 	"testing"
-	"time"
 )
 
 func TestIoLog(t *testing.T) {
@@ -113,30 +112,6 @@ func TestBufferLongLine(t *testing.T) {
 	logger.Writer = &nopcloser{buf}
 
 	logger.Log(Data{"a": 1})
-	expected := "a=1\n"
-	if actual := buf.String(); actual != expected {
-		t.Errorf("e: %s\na: %s", strconv.Quote(expected), strconv.Quote(actual))
-	}
-}
-
-func TestBufferTimeout(t *testing.T) {
-	buf := bytes.NewBufferString("")
-	logger := NewBufferedLogger("", 10)
-	logger.AddTime = false
-	logger.Writer = &nopcloser{buf}
-	logger.Timeout = 1 * time.Millisecond
-
-	ch := make(chan Data, 3)
-	ch <- Data{"a": 1}
-
-	go logger.Watch(ch)
-
-	time.Sleep(10 * time.Millisecond)
-	if logger.Position != 0 {
-		t.Errorf("Not flushed: %d", logger.Position)
-	}
-	ch <- nil
-
 	expected := "a=1\n"
 	if actual := buf.String(); actual != expected {
 		t.Errorf("e: %s\na: %s", strconv.Quote(expected), strconv.Quote(actual))
